audio

+++ hasm/src/hasm.c
@@ -37,7 +37,7 @@ uint32_t Hasm_Assemble(Hasm* me, const char* ifilename, uint8_t* ram, int startA
 	me->baseDir = calloc(1, GetDir(ifilename, NULL));
 	GetDir(ifilename, me->baseDir);
 
-	uint16_t ret = Assemble(me, ifilename, startAddr, 0);
+	uint32_t ret = Assemble(me, ifilename, startAddr, 0);
 	Labels_Replace(me->labels, me->ram);
 
 	return ret;
